Fixes #6 removes hard coded sitemap.xml
Instead of hard coding sitemap.xml (everyone may not have this, and
everyone may not want to move it to the asset folder). Instead, use the
new `additionalFiles` option.

NOTE: the sitemap.xml was copied, not moved. Using the `additionalFiles`
option will move the file instead of copying it. Is that ok?
I don't know why you would rather copy it than moving it... Need to ask
the maintainer of this plugin to know if this is ok.