Using additional files (#18)
* Adds code style to default arg in readme

* Fixes #12 removes move of manifest

Instead of hard coding that the manifest should be moved, use the new
`additionalFiles` option.

* Fixes #6 removes hard coded sitemap.xml

Instead of hard coding sitemap.xml (everyone may not have this, and
everyone may not want to move it to the asset folder). Instead, use the
new `additionalFiles` option.

NOTE: the sitemap.xml was copied, not moved. Using the `additionalFiles`
option will move the file instead of copying it. Is that ok?
I don't know why you would rather copy it than moving it... Need to ask
the maintainer of this plugin to know if this is ok.

* Updates readme with breaking changed in v2